Remove NewsArticle schema and add dynamic page schemas
